完善波形控制，完善故障设置后台通信

+++ app/src/main/java/com/example/autosimandroid/WaveView.java
@@ -58,7 +58,7 @@ public class WaveView extends ImageView {
         channelHashMap.get(channel).wave = wave;
     }
 
-    public void setPos(String channel, int pos) {
+    public void setPos(String channel, float pos) {
         channelHashMap.get(channel).pos = pos;
     }
